Add e2e setup
Adds resources to test a mapping configuration for kubernetes.
It attempts to solve the problem of a shared index between applications
that can cause mapping conflicts which lead to buffer saturation and
logging outages.

It uses Kubernetes in Docker (KinD) to startup the cluster and run the
locally built image with the custom configuration.
It deploys a pod that will log entries which would cause a conflict
mapping, and it would be expected that log-shipper will have no issues.

On a following PR the tests will be implemented to make sure that the
logs are being sent to the appropriate index.